# docs/swagger.yaml
basePath: /
definitions:
...
  request.CustomerCreateRequest:
    properties:
      address:
        type: string
      coa_id:
        type: string
      code:
        maxLength: 20
        minLength: 1
        type: string
      email:
        type: string
      name:
        minLength: 1
        type: string
      phone:
        type: string
    required:
    - code
    - name
    type: object
  request.CustomerUpdateRequest:
    properties:
      address:
        type: string
      coa_id:
        type: string
      code:
        maxLength: 20
        minLength: 1
        type: string
      email:
        type: string
      name:
        minLength: 1
        type: string
      phone:
        type: string
    required:
    - code
    - name
    type: object
  request.JournalEntryCreateRequest:
    properties:
      date:
        type: string
      description:
        minLength: 1
        type: string
      lines:
        items:
          $ref: '#/definitions/request.JournalLineRequest'
        minItems: 2
        type: array
    required:
    - date
    - description
    - lines
    type: object
  request.JournalEntryUpdateRequest:
    properties:
      date:
        type: string
      description:
        minLength: 1
        type: string
      lines:
        items:
          $ref: '#/definitions/request.JournalLineRequest'
        minItems: 2
        type: array
    required:
    - date
    - description
    - lines
    type: object
  request.JournalLineRequest:
    properties:
      coa_id:
        type: string
      credit:
        minimum: 0
        type: number
      debit:
        minimum: 0
        type: number
      description:
        minLength: 1
        type: string
    required:
    - coa_id
    - description
    type: object
...
  response.JSON:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/response.Meta'
      status:
        type: integer
    type: object
  response.Links:
    properties:
      current_page:
        type: string
      first_page:
        type: string
      last_page:
        type: string
      next_page:
        type: string
      prev_page:
        type: string
    type: object
  response.Meta:
    properties:
      info:
        type: string
      links:
        $ref: '#/definitions/response.Links'
      page:
        type: integer
      page_size:
        type: integer
      search:
        type: string
      total_count:
        type: integer
      total_pages:
        type: integer
    type: object
  response.SelectJSON:
    properties:
      data: {}
    type: object
...
host: 127.0.0.1:3000
info:
  contact:
    email: fatihrizqon@gmail.com
    name: Fatih Rizqon
  description: This is an Official Documentation for Go REST API with Fiber Framework
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go REST API with Fiber Framework
  version: "1.0"
paths:
...
  /api/v1/customers:
    get:
      consumes:
      - application/json
      description: Retrieve all customer records with pagination
      parameters:
      - description: Search keyword
        in: query
        name: search
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Sort column (code, name, email, status, created_at, updated_at)
        in: query
        name: sort
        type: string
      - description: Sort direction (asc, desc)
        in: query
        name: order
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved all records.
          schema:
            $ref: '#/definitions/response.JSON'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSON'
      security:
      - BearerAuth: []
      summary: Get all customers
      tags:
      - Customers
    post:
      consumes:
      - application/json
      description: Store a new customer record
      parameters:
      - description: Customer Create Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CustomerCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: A new record has been stored.
          schema:
            $ref: '#/definitions/response.JSON'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.JSON'
      security:
      - BearerAuth: []
      summary: Create customer
      tags:
      - Customers
  /api/v1/customers/{id}:
    delete:
      consumes:
      - application/json
      description: Remove a customer record by ID
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Selected record has been deleted.
          schema:
            $ref: '#/definitions/response.JSON'
        "404":
          description: Customer not found
          schema:
            $ref: '#/definitions/response.JSON'
      security:
      - BearerAuth: []
      summary: Delete customer
      tags:
      - Customers
    get:
      consumes:
      - application/json
      description: Retrieve a single customer by its ID
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved selected record.
          schema:
            $ref: '#/definitions/response.JSON'
        "404":
          description: Customer not found
          schema:
            $ref: '#/definitions/response.JSON'
      security:
      - BearerAuth: []
      summary: Get customer by ID
      tags:
      - Customers
    put:
      consumes:
      - application/json
      description: Update customer data by ID
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: string
      - description: Customer Update Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CustomerUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Selected record has been updated.
          schema:
            $ref: '#/definitions/response.JSON'
        "404":
          description: Customer not found
          schema:
            $ref: '#/definitions/response.JSON'
      security:
      - BearerAuth: []
      summary: Update customer
      tags:
      - Customers
...
  /api/v1/dropdown/customers:
    get:
      consumes:
      - application/json
      description: Retrieve customers for dropdown selection
      parameters:
      - description: Search keyword
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved dropdown options
          schema:
            $ref: '#/definitions/response.SelectJSON'
      security:
      - BearerAuth: []
      summary: Get customer dropdown options
      tags:
      - Dropdowns
...
  /api/v1/expenses:
    get:
      parameters:
      - description: Search keyword
        in: query
        name: search
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Filter by status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSON'
      summary: Get all expense entries
      tags:
      - Expenses
    post:
      consumes:
      - application/json
      description: Store a new expense journal entry (type=expense, prefix EX-).
      parameters:
      - description: Expense Create Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.JournalEntryCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.JSON'
      security:
      - BearerAuth: []
      summary: Create expense entry
      tags:
      - Expenses
  /api/v1/expenses/{id}:
    delete:
      parameters:
      - description: Entry ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSON'
      summary: Delete expense entry
      tags:
      - Expenses
    get:
      parameters:
      - description: Entry ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSON'
      summary: Get expense entry by ID
      tags:
      - Expenses
    put:
      consumes:
      - application/json
      parameters:
      - description: Entry ID
        in: path
        name: id
        required: true
        type: string
      - description: Expense Update Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.JournalEntryUpdateRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSON'
      security:
      - BearerAuth: []
      summary: Update expense entry
      tags:
      - Expenses
  /api/v1/expenses/{id}/post:
    put:
      parameters:
      - description: Entry ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSON'
      summary: Post expense entry
      tags:
      - Expenses
  /api/v1/expenses/{id}/void:
    put:
      parameters:
      - description: Entry ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSON'
      summary: Void expense entry
      tags:
      - Expenses
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"